+++ src/util/artifacts.rs
@@ -33,7 +33,7 @@ where
         let version = artifact.version();
 
         log::info!("installing artifact {}", &version);
-        dxm_artifacts::install(&client, &platform, &version, artifact.path(&path))?;
+        dxm_artifacts::install(&client, &platform, version, artifact.path(&path))?;
 
         Lockfile::write_artifact_version(path, version)?;